Install a Qt3D widget in the viewer area
This removes the QOpenGLWidget from the .ui file.  It leaves
the ui editor a bit confused about what will be using up
space or not; there's no good way that I know of to explain
to the editor about QWidget.createWindowContainer, which
we must use to insert a Qt3D panel into a QWidgets window.